Finalização da implementação dos testes unitários.

+++ src/main/java/br/com/fiap/tech/challenge/grupo/vinte/oito/ChefOnline/core/dtos/cardapio/ItemCardapioDTO.java
@@ -2,13 +2,14 @@ package br.com.fiap.tech.challenge.grupo.vinte.oito.ChefOnline.core.dtos.cardapi
 import br.com.fiap.tech.challenge.grupo.vinte.oito.ChefOnline.core.domain.entities.cardapio.DisponibilidadeConsumoPedidoEnum;
 import br.com.fiap.tech.challenge.grupo.vinte.oito.ChefOnline.core.dtos.restaurante.RestauranteDTO;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 public record ItemCardapioDTO(
     Long id,
     String nome,
     String descricao,
-    Double preco,
+    BigDecimal preco,
     List<DisponibilidadeConsumoPedidoEnum> disponibilidadeConsumo,
     String foto,
     RestauranteDTO restaurante
